etl/requirements.txt
google-cloud-bigquery==3.*
google-cloud-storage==2.*
pyarrow>=15
